refactor(core): Protocol composes internal Dispatcher+StreamDriver; dispatch() public (R2 of decomposition; behavior unchanged)
